backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
pgvector
python-dotenv
requests
